[REFACTOR] Service 리팩토링

+++ src/main/java/com/holly/molly/service/ReviewService.java
@@ -18,7 +18,7 @@ import java.util.Optional;
 public class ReviewService {
     private final ReviewRepository reviewRepository;
 
-    @Transactional//필요시에만 readOnly=false
+    @Transactional
     public Long join(Review review){
         reviewRepository.save(review);
         return review.getId();
@@ -26,11 +26,14 @@ public class ReviewService {
 
     @Transactional
     public Long delete(Review review){
+        if(Optional.ofNullable(reviewRepository.findOne(review.getId())).isEmpty()){
+            throw new RuntimeException("삭제하려는 Accept가 존재하지 않습니다.");
+        }
         reviewRepository.delete(review);
         return review.getId();
     }
 
-    public Review findOne(Long id){ return reviewRepository.findOne(id); }
+    public Optional<Review> findOne(Long id){ return Optional.ofNullable(reviewRepository.findOne(id)); }
 
     public List<Review> findAll(){
         return reviewRepository.findAll();
